# backend/docs/swagger.yaml
definitions:
...
  domain.ChunkListItemResp:
    properties:
      content:
        type: string
      id:
        type: string
      seq:
        type: integer
      title:
        type: string
    type: object
...
  domain.CreateDocReq:
    properties:
      content:
        type: string
      file_key:
        items:
          type: string
        type: array
      kb_id:
        type: string
      source:
        allOf:
        - $ref: '#/definitions/domain.DocSource'
        enum:
        - 1
        - 2
        - 3
      title:
        type: string
      url:
        items:
          type: string
        type: array
    required:
    - kb_id
    - source
    type: object
...
  domain.DocActionReq:
    properties:
      action:
        enum:
        - delete
        type: string
      doc_ids:
        items:
          type: string
        type: array
    required:
    - action
    - doc_ids
    type: object
  domain.DocDetailResp:
    properties:
      content:
        type: string
      created_at:
        type: string
      error:
        type: string
      id:
        type: string
      meta:
        $ref: '#/definitions/domain.DocMeta'
      resource_type:
        $ref: '#/definitions/domain.ResourceType'
      source:
        $ref: '#/definitions/domain.DocSource'
      status:
        $ref: '#/definitions/domain.DocStatus'
      updated_at:
        type: string
      url:
        type: string
    type: object
  domain.DocListItemResp:
    properties:
      chunk_count:
        type: integer
      created_at:
        type: string
      error:
        type: string
      favicon:
        type: string
      id:
        type: string
      resource_type:
        $ref: '#/definitions/domain.ResourceType'
      source:
        $ref: '#/definitions/domain.DocSource'
      status:
        $ref: '#/definitions/domain.DocStatus'
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
      word_count:
        type: integer
    type: object
  domain.DocMeta:
    properties:
      charset:
        type: string
      description:
        type: string
      doc_type:
        $ref: '#/definitions/domain.ResourceType'
      favicon:
        type: string
      keywords:
        type: string
      screenshot:
        type: string
      title:
        type: string
    type: object
  domain.DocSource:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - DocSourceURL
    - DocSourceFile
    - DocSourceManual
  domain.DocStatus:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - DocStatusPending
    - DocStatusPublished
    - DocStatusFailed
...
  domain.ObjectUploadResp:
    properties:
      key:
        type: string
    type: object
  domain.ParseURLItem:
    properties:
      desc:
        type: string
      published:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  domain.ParseURLReq:
    properties:
      type:
        enum:
        - RSS
        - Sitemap
        type: string
      url:
        type: string
    required:
    - type
    - url
    type: object
  domain.ParseURLResp:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.ParseURLItem'
        type: array
    type: object
...
  domain.ResourceType:
    enum:
    - WebPage
    - PlainText
    - Excel
    - Word
    - PDF
    - Unknown
    type: string
    x-enum-varnames:
    - ResourceTypeWebPage
    - ResourceTypePlainText
    - ResourceTypeExcel
    - ResourceTypeWord
    - ResourceTypePDF
    - ResourceTypeUnknown
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.UpdateDocReq:
    properties:
      content:
        type: string
      doc_id:
        type: string
      title:
        type: string
    required:
    - doc_id
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/doc:
    post:
      consumes:
      - application/json
      description: Create Doc
      parameters:
      - description: Doc
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDocReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Create Doc
      tags:
      - doc
  /api/v1/doc/action:
    post:
      consumes:
      - application/json
      description: Doc Action
      parameters:
      - description: Action
        in: body
        name: action
        required: true
        schema:
          $ref: '#/definitions/domain.DocActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Doc Action
      tags:
      - doc
  /api/v1/doc/chunk/list:
    get:
      consumes:
      - application/json
      description: Get Chunk List
      parameters:
      - description: Doc ID
        in: query
        name: doc_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ChunkListItemResp'
                  type: array
              type: object
      summary: Get Chunk List
      tags:
      - doc
  /api/v1/doc/detail:
    get:
      consumes:
      - application/json
      description: Get Doc Detail
      parameters:
      - description: Doc ID
        in: query
        name: doc_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.DocDetailResp'
              type: object
      summary: Get Doc Detail
      tags:
      - doc
    put:
      consumes:
      - application/json
      description: Update Doc Detail
      parameters:
      - description: Doc
        in: body
        name: doc
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateDocReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Update Doc Detail
      tags:
      - doc
  /api/v1/doc/list:
    get:
      consumes:
      - application/json
      description: Get Doc List
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.DocListItemResp'
                  type: array
              type: object
      summary: Get Doc List
      tags:
      - doc
  /api/v1/doc/parse_url:
    post:
      consumes:
      - application/json
      description: Parse URL
      parameters:
      - description: Parse URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ParseURLReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ParseURLResp'
              type: object
      summary: Parse URL
      tags:
      - doc
  /api/v1/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload File
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: Knowledge Base ID
        in: formData
        name: kb_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ObjectUploadResp'
      summary: Upload File
      tags:
      - file
...
swagger: "2.0"